requirements.txt
django==3.1.2
djangorestframework==3.13.1
psycopg2==2.9.5
requests